feat: extract utility functions into @workspace/utils package
