+++ apps/vbcall/package.json
@@ -20,6 +20,7 @@
     "@t3-oss/env-nextjs": "0.13.8",
     "@tanstack/table-core": "8.21.3",
     "@workspace/ui": "workspace:*",
+    "@workspace/utils": "workspace:*",
     "@workspace/vbcall-db": "workspace:*",
     "aws-amplify": "6.15.3",
     "babel-plugin-react-compiler": "19.1.0-rc.2",
