+++ apps/vbpay/app/(authed)/network/entity/[slug]/hooks/use-edit-entity-form.tsx
@@ -7,7 +7,7 @@ import { useForm } from "react-hook-form";
 import { toast } from "sonner";
 
 import { useErrorDialog } from "@workspace/ui/hooks/use-error-dialog";
-import { getErrorMessage } from "@workspace/ui/lib/get-error-message";
+import { getErrorMessage } from "@workspace/utils/get-error-message";
 
 import { UserType } from "@/types/user-type";
