+++ apps/vbpay/app/(authed)/network/physician/[slug]/components/info/edit-physician-form/edit-physician-form.tsx
@@ -8,7 +8,7 @@ import { FormInput } from "@workspace/ui/components/form/form-input";
 import { FormRadioGroup } from "@workspace/ui/components/form/form-radio-group";
 import { FormSubmitButton } from "@workspace/ui/components/form/form-submit-button";
 import { ScrollArea } from "@workspace/ui/components/scroll-area";
-import { formatTaxId } from "@workspace/ui/lib/formatTaxId";
+import { formatTaxId } from "@workspace/utils/format-tax-id";
 
 import {
   NetworkPhysicianClasses,
