+++ apps/vbpay/app/(authed)/network/physicians/components/add-network-physician-form/add-network-physician-form-schema.ts
@@ -1,6 +1,6 @@
 import * as z from "zod";
 
-import { validateStringIsNumbers } from "@workspace/ui/lib/validateStringIsNumbers";
+import { validateStringIsNumbers } from "@workspace/utils/validate-string-is-numbers";
 
 import { NetworkPhysicianType } from "@/types/network-physician-type";
