+++ apps/vbpay/components/nppes-network-physician-search-form/nppes-network-physician-search-form-schema.ts
@@ -1,6 +1,6 @@
 import * as z from "zod";
 
-import { validateStringIsNumbers } from "@workspace/ui/lib/validateStringIsNumbers";
+import { validateStringIsNumbers } from "@workspace/utils/validate-string-is-numbers";
 
 const NppesNetworkPhysicianSearchFormSchema = z.object({
   npi: z
