+++ apps/vbpay/components/nppes-network-physician-table/nppes-network-physician-table-columns.tsx
@@ -8,8 +8,8 @@ import { ColumnDef } from "@tanstack/table-core";
 import { Button } from "@workspace/ui/components/button";
 import { ClientFormattedDate } from "@workspace/ui/components/client-formatted-date";
 import { DataTableColumnHeader } from "@workspace/ui/components/data-table/data-table-column-header";
-import { formatDate } from "@workspace/ui/lib/format-date";
-import { stringToTitleCase } from "@workspace/ui/lib/stringToTitleCase";
+import { formatDate } from "@workspace/utils/format-date";
+import { stringToTitleCase } from "@workspace/utils/string-to-title-case";
 
 import { NppesApiResponseResult } from "@/types/nppes-api-reponse";
 import { NppesSoleProprietor } from "@/types/nppes-sole-proprietor";
