+++ package.json
@@ -25,7 +25,7 @@
     "prettier-plugin-prisma": "5.0.0",
     "prettier-plugin-tailwindcss": "0.6.14",
     "sst": "3.17.10",
-    "turbo": "2.5.4",
+    "turbo": "2.5.5",
     "typescript": "5.8.3"
   },
   "packageManager": "pnpm@10.13.1",
