+++ packages/eslint-config/package.json
@@ -19,7 +19,7 @@
     "eslint-plugin-react": "7.37.5",
     "eslint-plugin-react-compiler": "19.1.0-rc.2",
     "eslint-plugin-react-hooks": "5.2.0",
-    "eslint-plugin-turbo": "2.5.4",
+    "eslint-plugin-turbo": "2.5.5",
     "globals": "16.3.0",
     "typescript": "5.8.3",
     "typescript-eslint": "8.37.0"
