+++ packages/ui/package.json
@@ -30,6 +30,7 @@
     "@tanstack/match-sorter-utils": "8.19.4",
     "@tanstack/react-table": "8.21.3",
     "@tanstack/table-core": "8.21.3",
+    "@workspace/utils": "workspace:*",
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
     "cmdk": "1.1.1",
@@ -47,7 +48,7 @@
   },
   "devDependencies": {
     "@tailwindcss/postcss": "4.1.11",
-    "@turbo/gen": "2.5.4",
+    "@turbo/gen": "2.5.5",
     "@types/node": "24.0.14",
     "@types/qrcode": "1.5.5",
     "@types/react": "19.1.8",
