+++ packages/ui/src/hooks/auth/use-mfa-verification-form.ts
@@ -2,13 +2,14 @@ import { useState } from "react";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 
+import { getErrorMessage } from "@workspace/utils/get-error-message";
+
 import { handleSignInNextStep } from "../../lib/auth/handle-sign-in-next-step";
 import {
   MfaVerificationFormDefaultValues,
   MfaVerificationFormOutput,
   MfaVerificationFormSchema,
 } from "../../lib/auth/mfa-verification-form-schema";
-import { getErrorMessage } from "../../lib/get-error-message";
 import type { ConfirmSignInFunction, SignInResult } from "../../types/auth";
 import { useErrorDialog } from "../use-error-dialog";
