+++ packages/ui/src/hooks/auth/use-sign-in-form.ts
@@ -2,13 +2,14 @@ import { useState } from "react";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 
+import { getErrorMessage } from "@workspace/utils/get-error-message";
+
 import { handleSignInNextStep } from "../../lib/auth/handle-sign-in-next-step";
 import {
   SignInFormDefaultValues,
   SignInFormOutput,
   SignInFormSchema,
 } from "../../lib/auth/sign-in-form-schema";
-import { getErrorMessage } from "../../lib/get-error-message";
 import type { SignInFunction, SignInResult } from "../../types/auth";
 import { useErrorDialog } from "../use-error-dialog";
